webapp/requirements.txt
Django==4.2.15
django-htmx==1.19.0
psycopg2-binary==2.9.9
gunicorn==22.0.0
Pillow==10.4.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
django-ckeditor==6.7.1